oj-web/requirements.txt
Django==4.2.4
celery==5.2.7
redis==4.3.4
django-cors-headers==4.2.0
django-import-export==3.3.6
djangorestframework==3.14.0
#uwsgi
gunicorn==23.0.0
openpyxl==3.1.2
requests==2.31.0
markdown==3.7
# for postgres db
# psycopg2-binary
psycopg[binary]==3.1.12
psycopg[c]